Complete premium dashboard redesign: extract tabs, remove school colors, unify formatting
Decompose monolithic app.py (2573 lines → 181 lines) into modular architecture:

- Extract 7 tab rendering functions into tabs/ module (tiger5, strokes_gained,
  driving, approach, short_game, putting, coachs_corner)
- Replace all ODU school colors (gold/black/red) with neutral premium palette
  (charcoal/blue/indigo) from centralized ui/theme.py design tokens
- Replace all inline HTML card markup with reusable components:
  PremiumHeroCard, PremiumStatCard, SectionHeader
- Unify conditional formatting: all thresholds centralized in THRESHOLDS dict,
  same semantic colors (green=positive, red=negative) across all 7 tabs
- Centralize date/SG/percentage formatting in engine files via
  round_label() and format_date() from ui/formatters.py
- Update chart colors to use CHART_PALETTE from theme tokens
- Remove duplicate _sg_card_class() definitions (was in short_game + putting)
