+++ engines/driving.py
@@ -1,5 +1,6 @@
 import pandas as pd
 import numpy as np
+from ui.formatters import round_label
 
 # ============================================================
 # DRIVING ENGINE
@@ -298,7 +299,7 @@ def build_driving_results(filtered_df, num_rounds, hole_summary):
     round_trend = round_trend.sort_values('Date')
     round_trend['Fairway %'] = round_trend['Fairway_Count'] / round_trend['Total_Drives'] * 100
     round_trend['Label'] = round_trend.apply(
-        lambda r: f"{r['Date'].strftime('%m/%d/%y')} {r['Course']}", axis=1
+        lambda r: round_label(r['Date'], r['Course']), axis=1
     )
 
     return {
